week1/requirements.txt
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
jupyterlab>=4.0.0,<5.0.0
ipykernel>=6.25.0
